Return ID of sent message
Sending a message will now return its new serialized id so you can, for exemple, save on a database to use later with ack.
As of the official API, a successful response includes an identifier for the newly created message. (Sending Text Messages)

Note: WAPI.sendMessage will still return the whole object so we don't mess with pre-existing codes using the function like this, but in my opinion it should return only the id aswell.